Move StratumServerClient to its own file
