+++ ironfish/src/mining/pool.ts
@@ -14,7 +14,8 @@ import { ErrorUtils } from '../utils/error'
 import { FileUtils } from '../utils/file'
 import { SetTimeoutToken } from '../utils/types'
 import { MiningPoolShares } from './poolShares'
-import { StratumServer, StratumServerClient } from './stratum/stratumServer'
+import { StratumServer } from './stratum/stratumServer'
+import { StratumServerClient } from './stratum/stratumServerClient'
 import { mineableHeaderString } from './utils'
 import { WebhookNotifier } from './webhooks'
