+++ ironfish/src/mining/stratum/errors.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import * as yup from 'yup'
-import { StratumServerClient } from './stratumServer'
+import { StratumServerClient } from './stratumServerClient'
 
 export class MessageMalformedError extends Error {
   constructor(sender: string, error: yup.ValidationError | string, method?: string) {
